hash finished
