+++ integrate/yigconf/yig.toml
@@ -21,7 +21,7 @@ iam_key = "key"
 iam_secret = "secret"
 
 # Meta Config
-meta_cache_type = 2
+meta_cache_type = 2 
 meta_store = "tidb"
 tidb_info = "root:@tcp(10.5.0.17:4000)/yig"
 keepalive = true
